requirements.txt
httpx==0.28.1
python-dotenv==1.1.0
python-telegram-bot==21.10
yt_dlp==2025.04.30
gallery-dl==1.29.6